[State Sync] Move logic into different files.
